@@ -2,9 +2,12 @@ package com.location.reminder.sound.db
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
+import androidx.room.TypeConverters
 import com.location.reminder.sound.model.Task
+import com.location.reminder.sound.util.Converters
 
-@Database(entities = [Task::class], version = 6)
+@Database(entities = [Task::class], version = 1, exportSchema = false)
+@TypeConverters(Converters::class)
 abstract class AppDatabase : RoomDatabase() {
     abstract fun taskDao(): TaskDao
 }
