@@ -9,12 +9,13 @@ import com.google.android.libraries.places.api.net.PlacesClient
 import com.location.reminder.sound.BuildConfig
 import com.location.reminder.sound.db.AppDatabase
 import com.location.reminder.sound.db.TaskDao
-import com.location.reminder.sound.repositories.HomeRepository
-import com.location.reminder.sound.repositories.HomeRepositoryImpl
+import com.location.reminder.sound.domain.repositories.HomeRepository
+import com.location.reminder.sound.domain.repositories.HomeRepositoryImpl
 import com.location.reminder.sound.location.LocationClient
 import com.location.reminder.sound.location.LocationClientImpl
 import com.location.reminder.sound.network.PlacesApi
 import com.location.reminder.sound.network.RetrofitUtil
+import com.location.reminder.sound.util.Constants
 import com.location.reminder.sound.util.SharedPrefClient
 import dagger.Module
 import dagger.Provides
@@ -28,52 +29,39 @@ import retrofit2.Retrofit
 class AppModule {
 
     @Provides
-    fun provideBaseUrl() = "Hi"
-
-    @Provides
-    fun provideSharedPreference(@ApplicationContext context: Context): SharedPrefClient {
-        return SharedPrefClient(context)
-    }
+    fun provideSharedPreference(@ApplicationContext context: Context): SharedPrefClient =
+        SharedPrefClient(context)
 
     @Provides
     fun getHomeRepository(
         sharedPrefClient: SharedPrefClient,
         database: TaskDao
-    ): HomeRepository {
-        return HomeRepositoryImpl(sharedPrefClient, database)
-    }
+    ): HomeRepository = HomeRepositoryImpl(sharedPrefClient, database)
 
     @Provides
     fun getLocationClient(
         @ApplicationContext context: Context,
         fusedLocationProviderClient: FusedLocationProviderClient
-    ): LocationClient {
-        return LocationClientImpl(context, fusedLocationProviderClient)
-    }
+    ): LocationClient = LocationClientImpl(context, fusedLocationProviderClient)
 
     @Provides
-    fun getFusedLocationProviderClient(@ApplicationContext context: Context): FusedLocationProviderClient {
-        return LocationServices.getFusedLocationProviderClient(context)
-    }
+    fun getFusedLocationProviderClient(@ApplicationContext context: Context): FusedLocationProviderClient =
+        LocationServices.getFusedLocationProviderClient(context)
 
     @Provides
     fun getDataBase(@ApplicationContext context: Context): TaskDao {
         return Room.databaseBuilder(
             context,
-            AppDatabase::class.java, "tasks"
+            AppDatabase::class.java, Constants.DB_NAME
         ).fallbackToDestructiveMigration().build().taskDao()
 
     }
 
     @Provides
-    fun getRetrofit(): Retrofit {
-        return RetrofitUtil.getInstance();
-    }
+    fun getRetrofit(): Retrofit = RetrofitUtil.getInstance()
 
     @Provides
-    fun getPlacesApi(retrofit: Retrofit): PlacesApi {
-        return retrofit.create(PlacesApi::class.java);
-    }
+    fun getPlacesApi(retrofit: Retrofit): PlacesApi = retrofit.create(PlacesApi::class.java)
 
     @Provides
     fun getPlacesClient(@ApplicationContext context: Context): PlacesClient {
